Use typeutil.Undefined in models and response DTOs
https://github.com/go-goyave/goyave/pull/244

It showcases the new ability to properly handle unselected columns and perfectly handle goyave.dev/filter "fields" feature, that only selects and returns the desired fields.

+++ database/repository/user.go
@@ -20,7 +20,7 @@ func NewUser(db *gorm.DB) *User {
 	}
 }
 
-func (r *User) GetByID(ctx context.Context, id uint) (*model.User, error) {
+func (r *User) GetByID(ctx context.Context, id int64) (*model.User, error) {
 	var user *model.User
 	db := session.DB(ctx, r.DB).Where("id", id).First(&user)
 	return user, errors.New(db.Error)
